Getting node status every 15 minutes

+++ netsnapshot/snapshot.go
@@ -48,6 +48,19 @@ func (t taker) Start(ctx context.Context) {
 		os.Exit(1)
 	}
 
+	// update all reachable nodes
+	loadLiveNodes := func() {
+		nodes, err := t.dataStore.GetAvailableNodes(ctx)
+		if err != nil {
+			log.Errorf("Error in taking network snapshot, %s", err.Error())
+		}
+		amgr.setLiveNodes(nodes)
+		log.Infof("%d nodes loaded to live reload", len(nodes))
+	}
+
+	// enqueue previous known ips
+	loadLiveNodes()
+
 	go runSeeder(t.cfg, netParams)
 
 	var mtx sync.Mutex
@@ -89,7 +102,6 @@ func (t taker) Start(ctx context.Context) {
 			})
 
 			if err != nil {
-				// todo delete all the related node info
 				t.dataStore.DeleteSnapshot(ctx, timestamp)
 				log.Errorf("Error in saving network snapshot, %s", err.Error())
 			}
@@ -99,6 +111,9 @@ func (t taker) Start(ctx context.Context) {
 			log.Infof("Took a new network snapshot, recorded %d discoverable nodes.", count)
 			timestamp = time.Now().UTC().Unix()
 			mtx.Unlock()
+			// update all reachable nodes
+			loadLiveNodes()
+
 
 		case node := <-amgr.goodPeer:
 			if node.IP.String() == "127.0.0.1" { // do not add the local IP
@@ -117,45 +132,63 @@ func (t taker) Start(ctx context.Context) {
 				CurrentHeight:   node.CurrentHeight,
 				Services: 		 node.Services.String(),
 				Latency: 		 int(node.Latency),
-
 			}
 
-			geoLoc, err := t.geolocation(ctx, node.IP)
-			if err == nil {
-				networkPeer.IPInfo = *geoLoc
-				// networkPeer.Country = geoLoc.CountryName
-				if geoLoc.Type == "ipv4" {
-					networkPeer.IPVersion = 4
-				} else if geoLoc.Type == "ipv6" {
-					networkPeer.IPVersion = 6
-				}
-
-			}
-
-			err = t.dataStore.SaveNetworkPeer(ctx, networkPeer)
-			if err != nil {
-				log.Errorf("Error in saving node info, %s.", err.Error())
-			} else {
-				mtx.Lock()
-				count++
-				if node.CurrentHeight > bestBlockHeight {
-					bestBlockHeight = node.CurrentHeight
-				}
-
-				snapshot := SnapShot{
-					Timestamp: timestamp,
-					Height:    bestBlockHeight,
-				}
-
-				err = t.dataStore.SaveSnapshot(ctx, snapshot)
+			if exists, _ := t.dataStore.NodeExists(ctx, networkPeer.Address); exists {
+				err = t.dataStore.UpdateNode(ctx, networkPeer)
 				if err != nil {
-					// todo delete all the related node info
-					t.dataStore.DeleteSnapshot(ctx, timestamp)
-					log.Errorf("Error in saving network snapshot, %s", err.Error())
+					log.Errorf("Error in saving node info, %s.", err.Error())
+				}
+			} else {
+				geoLoc, err := t.geolocation(ctx, node.IP)
+				if err == nil {
+					networkPeer.IPInfo = *geoLoc
+					// networkPeer.Country = geoLoc.CountryName
+					if geoLoc.Type == "ipv4" {
+						networkPeer.IPVersion = 4
+					} else if geoLoc.Type == "ipv6" {
+						networkPeer.IPVersion = 6
+					}
 				}
 
-				mtx.Unlock()
-				log.Infof("New heartbeat recorded for node: %s, %s, %d", node.IP.String(), node.UserAgent, node.ProtocolVersion)
+				err = t.dataStore.SaveNode(ctx, networkPeer)
+				if err != nil {
+					log.Errorf("Error in saving node info, %s.", err.Error())
+				}
+			}
+
+			// if this node is reachable within the current timestamp, save heartbeat
+			if networkPeer.LastSuccess >= timestamp {
+				err = t.dataStore.SaveHeartbeat(ctx, Heartbeat{
+					Timestamp: timestamp,
+					Address:   node.IP.String(),
+					LastSeen:  node.LastSeen.UTC().Unix(),
+					Latency:   int(node.Latency),
+				})
+				if err != nil {
+					log.Errorf("Error in saving node info, %s.", err.Error())
+				} else {
+					mtx.Lock()
+					count++
+					if node.CurrentHeight > bestBlockHeight {
+						bestBlockHeight = node.CurrentHeight
+					}
+
+					snapshot := SnapShot{
+						Timestamp: timestamp,
+						Height:    bestBlockHeight,
+					}
+
+					err = t.dataStore.SaveSnapshot(ctx, snapshot)
+					if err != nil {
+						// todo delete all the related node info
+						t.dataStore.DeleteSnapshot(ctx, timestamp)
+						log.Errorf("Error in saving network snapshot, %s", err.Error())
+					}
+
+					mtx.Unlock()
+					log.Infof("New heartbeat recorded for node: %s, %s, %d", node.IP.String(), node.UserAgent, node.ProtocolVersion)
+				}
 			}
 		case <-ctx.Done():
 			log.Info("Shutting down network seeder")
@@ -166,18 +199,8 @@ func (t taker) Start(ctx context.Context) {
 }
 
 func (t taker) geolocation(ctx context.Context, ip net.IP) (*IPInfo, error) {
-	countryName, ipv, err := t.dataStore.GetIPLocation(ctx, ip.String())
-	if err == nil {
-		var v string
-		if ipv == 4 {
-			v = "ipv4"
-		} else if ipv == 6 {
-			v = "ipv6"
-		}
-		return &IPInfo{CountryName: countryName, Type: v}, nil
-	}
 	url := fmt.Sprintf("http://api.ipstack.com/%s?access_key=fcd33d8814206ce1f0a255a2204ad71e&format=1", ip.String())
 	var geo IPInfo
-	err = helpers.GetResponse(ctx, &http.Client{Timeout: 3 * time.Second}, url, &geo)
+	err := helpers.GetResponse(ctx, &http.Client{Timeout: 3 * time.Second}, url, &geo)
 	return &geo, err
 }
